\name{f0.ep}
\alias{f0.ep}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Frames list
}
\description{
Given a data frame and a lme4 like formula, the function returns the response vector and 
design matrices for fixed and random effects. Internal usage.
}
\usage{
f0.ep(data, formula)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
A data frame.
}
  \item{formula}{
An lme4 like formula.
}
}
\keyword{Frames}
