\name{fit.mglmm}
\alias{fit.mglmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fit.mglmm
}
\description{
Fitted values from a multivariate compound Poisson mixed model. Internal usage. See \link{fitted.cp}.
}
\usage{
fit.mglmm(mod, fixef, ranef)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
'mer' object.
}
  \item{fixef}{
Estimated fixed effects vector.
}
  \item{ranef}{
Random effects vector.
}
}
\keyword{Fitted values}
