\name{fitted.cp}
\alias{fitted.cp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fitted.cp
}
\description{
Fitted values from a multivariate compound Poisson mixed model.
}
\usage{
fitted.cp(mod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
Fitted multivariate compound Poisson model.
}
}
\details{
Returns the fitted values of the model.
}
\keyword{Fitted values}
