\name{format0CP}
\alias{format0CP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
format0CP
}
\description{
Organazing the fixed effects vectors from the fitted pairwise models. Internal usage.
}
\usage{
format0CP(mod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
Fitted 'mer' object.
}
}
