\name{format1CP}
\alias{format1CP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
format1CP
}
\description{
List with the averaged estimates for the fixed effects, variace components 
matrix, dispersion parameter and index parameter. Internal usage.
}
\usage{
format1CP(df.m, formula, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df.m}{
Data frame with averaged estimates.
}
  \item{formula}{
lme4 like formula.
}
  \item{data}{
Data.
}
}