\name{logLik.cp}
\alias{logLik.cp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
logLik.cp
}
\description{
Log-likelihood function and degrees of freedom of the fitted multivariate compound Poisson model.
}
\usage{
logLik.cp(mod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
Fitted multivariate compound Poisson model.
}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{logLik}{Log-likelihood value}
\item{df}{Degrees of freedom}
%% ...
}