\name{mglmmCP}
\alias{mglmmCP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mglmmCP
}
\description{
Function that fits multivariate compound Poisson models using the pairwise approach 
outlined in Fieuws et. al (2006).
}
\usage{
mglmmCP(formula, id, data, cl = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{
Multivariate formula.
}
  \item{id}{
Factor that identifies the multivariate responses.
}
  \item{data}{
Multivariate data as generated by data.sim().
}
  \item{cl}{
Cluster to be used.
}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{fixef}{Estimated fixed effects.}
\item{VarCov}{Estimated variance components matrix.}
\item{phi}{Estimated dispersion parameter.} 
\item{p}{Estimated index parameter of the copmpound Poisson distribution.}
\item{logLik}{Log-likelihood of the model.}
\item{df}{Degrees of freedom of the model.}
\item{ranef}{Matrix of random effects.}
\item{fitted}{Fitted values.}
\item{residuals}{Pearson like residuals.}
%% ...
}
\references{
Fieuws, S.; Verbeke, G.; Boen, F; Delecluse, C. High dimensional multivariate mixed models for binary questionnaire data. Applied Statistics 55(4) (2006). 449-460.
}

\examples{
library(pair.mglmm)
library(powell)


# Simulated data
phi <- 1; p <- 1.6 

mydat <- data.sim(3, 'CP', exp, xi=p, phi=phi)
dat <- mydat$Data

mainForm0 <- value~-1+variable:period+(-1+variable|ID)
mainForm1 <- value~-1+variable+(-1+variable|ID)


# Pairwise models
cl<-makeCluster(4)
registerDoParallel(cl)
clusterEvalQ(cl, library(pair.mglmm))

system.time(m0.1 <- mglmmCP(mainForm0, dat$variable, dat))
system.time(m1.1 <- mglmmCP(mainForm1, dat$variable, dat))

stopCluster(cl)


# Multivariate models
system.time(m0 <- cpglmm(mainForm0, data=dat))
system.time(m1 <- cpglmm(mainForm1, data=dat))



# Log-likelihood ratio
logLik(m0)-logLik(m1)
m0.1$logLik-m1.1$logLik


# Random effects
m0.1$ranef
ranef(m0)

m1.1$ranef
ranef(m1)


# Fitted values
fit0.1 <- m0.1$fitted
fit0 <- fitted(m0)
cbind(fit0, fit0.1)

fit1.1 <- m1.1$fitted
fit1 <- fitted(m1)
cbind(fit1, fit1.1)


# Model comparison
m0
summary.cp(m0.1)

m1
summary.cp(m1.1)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Multivariate models}
