\name{pair.mglmm-package}
\alias{pair.mglmm-package}
\alias{pair.mglmm}
\docType{package}
\title{
Pairwise fitting for multivariate GLMM
}
\description{
Pairwise fitting for multivariate GLMM
}
\details{
\tabular{ll}{
Package: \tab pair.mglmm\cr
Type: \tab Package\cr
Version: \tab 1.0-4\cr
Date: \tab 2012-11-21\cr
License: \tab GPL\cr
}
}
\author{
Rubem Ceratti

Maintainer: Rubem Ceratti <rubem_ceratti@yahoo.com.br>
}
\keyword{ package }
