\name{ranef.cp}
\alias{ranef.cp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ranef.cp
}
\description{
Random effects of a multivariate compound Poisson model.
}
\usage{
ranef.cp(mod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
Fitted multivariate compound Poisson model.
}
}