\name{resid.cp}
\alias{resid.cp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
resid.cp
}
\description{
Residuals of the multivariate compound Poisson model.
}
\usage{
resid.cp(mod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
Fitted multivariate compound Poisson model.
}
}
\details{
Pearson like residuals, analogous to those provided by lme4 and cplm.
}
