\name{resid.mglmm}
\alias{resid.mglmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
resid.mglmm
}
\description{
Residuals of the multivariate compound Poisson model. Internal usage. See \link{fitted.cp}.
}
\usage{
resid.mglmm(mod, fit, p)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod}{
'mer' object.
}
  \item{fit}{
Fitted values.
}
  \item{p}{
Estimated index parameter.
}
}