\name{uhat}
\alias{uhat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
uhat
}
\description{
Condictional modes of the random effects. Internal usage.
}
\usage{
uhat(r, beta, S, phi, p)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{r}{
See \link{cll}.
}
  \item{beta}{
See \link{cll}.
}
  \item{S}{
See \link{cll}.
}
  \item{phi}{
See \link{cll}.
}
  \item{p}{
See \link{cll}.
}
}